pyproject.toml
[project]
name = "investment-iq"
version = "1.0.0"
description = "InvestmentIQ MVAS - Multi-Agent Financial Intelligence System"
authors = [
    {name = "Group 2", email = "group2@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["agentic-ai", "multi-agent-system", "investment-analysis", "mcp", "context-engineering"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
    # Observability and Evaluation
    "langsmith>=0.1.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-huggingface>=0.0.1",
    # Vector Database for RAG
    "pinecone>=5.0.0",
    # Hugging Face Integration
    "huggingface-hub>=0.20.0",
    "sentence-transformers>=2.2.0",
    # Text Processing
    "tiktoken>=0.5.0",
    # Machine Learning
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    # HTTP Requests
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    # Async Support
    "asyncio>=3.4.3",
    # Scheduling (for services/scheduler.py)
    "apscheduler>=3.10.0",
    # Dashboard
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.4.1",
]

[project.urls]
Homepage = "https://github.com/group2/investment-iq"
Documentation = "https://github.com/group2/investment-iq#readme"
Repository = "https://github.com/group2/investment-iq"
Issues = "https://github.com/group2/investment-iq/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "config", "utils", "core", "evaluation", "services", "backtesting", "apps", "telemetry"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
